+++ crates/rome_js_formatter/tests/quick_test.rs
@@ -1,5 +1,5 @@
 use rome_formatter_test::check_reformat::CheckReformat;
-use rome_js_formatter::context::{JsFormatOptions, Semicolons};
+use rome_js_formatter::context::{JsFormatOptions, QuoteStyle, Semicolons};
 use rome_js_formatter::format_node;
 use rome_js_parser::parse;
 use rome_js_syntax::JsFileSource;
@@ -13,13 +13,19 @@ mod language {
 // use this test check if your snippet prints as you wish, without using a snapshot
 fn quick_test() {
     let src = r#"
-const foo = @deco class {
-  //
-};
+(
+  <div
+    attr=''
+    atrr={''}
+    />
+)
 "#;
     let syntax = JsFileSource::tsx();
     let tree = parse(src, syntax);
-    let options = JsFormatOptions::new(syntax).with_semicolons(Semicolons::AsNeeded);
+    let options = JsFormatOptions::new(syntax)
+        .with_semicolons(Semicolons::AsNeeded)
+        .with_quote_style(QuoteStyle::Double)
+        .with_jsx_quote_style(QuoteStyle::Single);
     let result = format_node(options.clone(), &tree.syntax())
         .unwrap()
         .print()