+++ crates/rome_js_formatter/tests/specs/js/module/import/bare_import.js
@@ -1,16 +1,16 @@
 import "very_long_import_very_long_import_very_long_import_very_long_import_very_long_import_very_long_import_very_long_import_";
-import "very_long_import_very_long_import_very_long_import_very_long_import_very_long_import_very_long" with { type :   "json"}
-import "short" with {
+import "very_long_import_very_long_import_very_long_import_very_long_import_very_long_import_very_long" assert { type :   "json"}
+import "short" assert {
 
         type :   "json"
 }
 
-import "very_long_import_very_long_import_very" with {
+import "very_long_import_very_long_import_very" assert {
     // something good is here
     "type": /****/ "json"
         }
 
-import "very_long_import_very_long_import_very" with {
+import "very_long_import_very_long_import_very" assert {
 		// something good is here
 		"type": /****/ "json",
 		"type2" /****/ : "json",