+++ crates/rome_js_parser/test_data/inline/ok/export_from_clause.js
@@ -4,6 +4,6 @@ export { default as a } from "b";
 export * from "a";
 export * as c from "b";
 export * as default from "b"
-export * from "mod" with { type: "json" }
+export * from "mod" assert { type: "json" }
 export type * from "types";
 export type * as types from "types";